+++ modules/apps/portal-search-elasticsearch7/portal-search-elasticsearch7-impl/src/main/java/com/liferay/portal/search/elasticsearch7/internal/index/CompanyIndexFactoryHelper.java
@@ -85,6 +85,8 @@ public class CompanyIndexFactoryHelper {
 		String indexName, IndicesClient indicesClient, long companyId,
 		boolean resetBothIndexNames) {
 
+		_executeIndexContributorsBeforeRemove(indexName);
+
 		DeleteIndexRequest deleteIndexRequest = new DeleteIndexRequest(
 			indexName);
 
@@ -230,11 +232,32 @@ public class CompanyIndexFactoryHelper {
 		}
 	}
 
+	private void _executeIndexContributorBeforeRemove(
+		IndexContributor indexContributor, String indexName) {
+
+		try {
+			indexContributor.onBeforeRemove(indexName);
+		}
+		catch (Throwable throwable) {
+			_log.error(
+				StringBundler.concat(
+					"Unable to apply contributor ", indexContributor,
+					" when removing index ", indexName),
+				throwable);
+		}
+	}
+
 	private void _executeIndexContributorsAfterCreate(String indexName) {
 		for (IndexContributor indexContributor :
 				_indexContributorServiceTrackerList) {
 
 			_executeIndexContributorAfterCreate(indexContributor, indexName);
+		}
+	}
+
+	private void _executeIndexContributorsBeforeRemove(String indexName) {
+		for (IndexContributor indexContributor : getIndexContributors()) {
+			_executeIndexContributorBeforeRemove(indexContributor, indexName);
 		}
 	}
